Maps and Pairs (#4)
Adding Map and Pair enables parsing+typechecking of all HCA/skylab workflows :champagne:

Wart with Pair: grammatical ambiguity of ".left" and ".right" getters leads to two separate code paths for handling them, depending on whether the accessed pair is an identifier or some other expression. It would be nice to eliminate this bifurcation in the future by ensuring "left" and "right" can't parse as an identifier part.