ForwardReference linter (#45)
Links each Expr.Ident to the Decl or Call it references, which is added during document typechecking by threading it through the type environment (as 'context').

This link makes it easy to ask whether the Expr.Ident lexically precedes the Decl/Call. It will also facilitate detection of cyclic dependencies and unused declarations.