UnusedDeclaration and UnusedCall linters (#46)
UnusedDeclaration: a (non-output) value is declared, but not subsequently used in the task/workflow.

UnusedCall: the outputs of a call within a workflow are neither outputs from the workflow, nor used by anything else in the workflow.

In the process of developing these, we realized that the task commands in many draft-2 WDLs were not being parsed as expected (specifically, `${ }` delimited placeholders were ignored within `<<< >>>` delimited commands, which are indeed not recognized as placeholders in WDL 1.0). To address this, we introduce a rudimentary way of formulating different grammar versions, and detect the version from the document header, if any. The mechanism will need more polishing.

Typecheck task runtime section.

Fix several bugs and omissions in the AST, Walker, etc.