rudimentary URI import (#63)
Downloading is punted to the library's user via an `import_uri` callback. It won't crawl relative paths in downloaded documents (spec is unclear whether this is allowed).

Also, documents containing workflows with incomplete calls can be imported, so long as their workflow is not actually called. (To use their tasks)

closes #17