struct literals may omit optional fields (#168)
An object literal typechecks with a struct type even if it omits some fields, if the omitted fields have optional types. At runtime, when the literal is coerced to the struct instance, these fields are filled in with null values. This should be done earlier during eval of the literal, in the future when there's a new struct literal syntax making the desired struct type known at eval time.

Closes #167