add symbolism for workflow Gather operations (#189)
When a value or call output inside a workflow scatter or conditional section is referenced by some identifier expression outside of that section, there's an implied operation to gather the array of values/outputs (or conditional value). Add a `Gather` node which sits in `WDL.Expr.Ident.referee` to symbolize this operation, which will be useful in scheduling workflow execution.