task runtime: switch to swarm for docker container scheduling (#225)
Swarm is built-in to docker and includes a useful layer of resource scheduling logic (e.g. waiting until desired number of CPUs are "available") which we won't need to reinvent for parallel task scheduling. It requires epsilon configuration to get going on the local host (`docker swarm init`, invoked automatically here).

Evaluates runtime.cpu and applies it to the container as both a reservation and a limit. Defaults to 1.