optional File? outputs from tasks (#259)
As in https://github.com/openwdl/wdl/pull/310 a task may provide an output filename that does not exist if the declared type is File?, in which case a None value will result. It's a bit tricky because the File? types may be nested inside compound values.

Incidental: preserve `Value.expr` references through value coercion