provide containers membership in invoking user's group (#286)
If a docker image has been baked with a USER directive to drop privileges to some arbitrary uid (a good security practice), then it may have trouble accessing the mounted working directory, which is owned by the user who invoked miniwdl. To support this we

1. provide the in-container command with membership in the invoking user's primary group
2. ensure everything miniwdl provisions in the working directory has group read/write permission

Incidentally add a gs:// downloader and a test script mirroring the "Getting Started" tutorial, which provoked the permissions change.