allow subworkflow name in input JSON key
To supply an optional input for a subworkflow call, Cromwell input JSON expects the key to include (i) the workflow name, (ii) the call name, (iii) the subworkflow name, and finally (iv) the input name. The need to include (iii) the subworkflow name is questionable, as discussed on https://github.com/openwdl/wdl/pull/340

This change makes miniwdl accept such inputs either with or without the subworkflow name, by removing it if it's present. To ensure unambiguous resolution, also forbids collision between a call's name and that of its containing workflow.

#193