config option to set hard memory limit from task runtime.memory (#365)
The task runtime.memory is used as a scheduling reservation, but isn't enforced as a hard limit. Here we add config option [task_runtime] memory_limit_multiplier, which leads to a hard limit of memory_limit_multiplier*runtime.memory if positive. It might be set to 1.0, below, or above depending on the circumstances.