Call caching -- first cut (#386)
#308 - call caching
#339 - digest code for inputs & tasks to inform call caching

Digestion - Use sha256 to create hash of the task and its inputs
Caching - Digest task and store json of outputs under taskName_task_digest/output_digest
Using the cache - When running a task check for cached outputs for that task/output set. If available return the stored outputs and move forward in the workflow without running that task.