address download cache race condition (#420)
Race condition: while we're downloading a URI, the intended cache location gets populated by some other download task for the same URI, possibly in another process; and it's being used as an input to that other run. When we finish our download, we overwrite the existing copy, potentially causing the other run to fail.

Now we decline to overwrite the existing copy, as long as [download_cache] get = true, i.e. it hadn't existed when we started the download. This makes the conflict very unlikely, although it's not eliminated completely. It also creates a possibility that large, redundant downloads can be left behind in the download cache 'ops' folder until it's cleaned up.