Make inputs with defaults implicitly optional (#507)
A call may supply None for an input with a non-optional type if it has a default expression, in which case the default will be used. In contrast, if the input has both a default expression and an optional type, None would be used. The WDL specification is ambiguous as to the latter case, though, so a new UnnecessaryQuantifier lint warning discourages relying on it.