Fix WDL.Env._rev with shadowed bindings
The result of WDL.Env._rev would include a shadowed binding, if one exists, *instead of* the visible one. This affected WDL.Env.merge and WDL.Env.Bindings.wrap_namespace.

The internal runner was unaffected because it does not rely on shadowing bindings, but, this was a serious bug for API consumers doing so. (issue #637 @adamnovak)